Allow both public and private key arguments for the Wallet object
The current SDK requires only a private key for creating and using
the wallet object, which is also used to verify the allowance.

This conflicts with the original idea for Paradigm that makers would
provide only the public key when adding their credentials to the admin
page, thus only the public key would be available when verifying the
allowance.

This can be solved with an alternative implementation:
* having both private and public key as optional arguments in the wallet
object
* raise an error when creating the wallet object if both are missing
* derive the public key from the private one only, if the private one
is available
* raise an error in sign_bid if the private key was not provided