[friktion] Adjust signature to include swap order addr
Adding the swap order address to the signature to ensure that it
is valid only for a given offer.

Also, adjust some functions to take the swap order addr as a
parameter to avoid extra calls to the blockchain.

WIP: this is not a final PR. It only antecipates some changes
Friktion will make to allow doing the necessary adjustments in the
Paradigm side